tools/ipc: explicitly cast send/recv return code
We should be using ssize_t for the code return but they are not
supported on Windows so we're instead explicit about the cast.

+++ ngl-tools/ipc.c
@@ -177,7 +177,7 @@ void ipc_pkt_freep(struct ipc_pkt **pktp)
 
 int ipc_send(int fd, const struct ipc_pkt *pkt)
 {
-    const int n = send(fd, pkt->data, pkt->size, 0);
+    const int n = (int)send(fd, pkt->data, pkt->size, 0);
     if (n < 0) {
         perror("send");
         return NGL_ERROR_IO;
@@ -194,7 +194,7 @@ static int readbuf(int fd, uint8_t *buf, int size)
 {
     int nr = 0;
     while (nr != size) {
-        const int n = recv(fd, buf + nr, size - nr, 0);
+        const int n = (int)recv(fd, buf + nr, size - nr, 0);
         if (n == 0)
             return 0;
         if (n < 0) {
